Move git_uptime() to l.lib.sysstats.
