+++ landscape/lib/process.py
@@ -2,12 +2,13 @@ import logging
 import os
 from datetime import timedelta, datetime
 
+from landscape.lib import sysstats
 from landscape.lib.timestamp import to_timestamp
 from landscape.lib.jiffies import detect_jiffies
 
 # FIXME: It'd be nice to avoid having library code which depends on
 #        landscape-specific modules.
-from landscape.monitor.computeruptime import BootTimes, get_uptime
+from landscape.monitor.computeruptime import BootTimes
 
 
 class ProcessInformation(object):
@@ -107,7 +108,7 @@ class ProcessInformation(object):
                 start_time = int(parts[21])
                 utime = int(parts[13])
                 stime = int(parts[14])
-                uptime = self._uptime or get_uptime()
+                uptime = self._uptime or sysstats.get_uptime()
                 pcpu = calculate_pcpu(utime, stime, uptime,
                                       start_time, self._jiffies_per_sec)
                 process_info["percent-cpu"] = pcpu
