+++ landscape/monitor/computeruptime.py
@@ -3,19 +3,9 @@ from datetime import datetime
 import os
 import struct
 
+from landscape.lib import sysstats
 from landscape.lib.timestamp import to_timestamp
 from landscape.monitor.plugin import MonitorPlugin
-
-
-def get_uptime(uptime_file=u"/proc/uptime"):
-    """
-    This parses a file in /proc/uptime format and returns a floating point
-    version of the first value (the actual uptime).
-    """
-    with open(uptime_file, 'r') as ufile:
-        data = ufile.readline()
-    up, idle = data.split()
-    return float(up)
 
 
 class LoginInfo(object):
@@ -111,7 +101,7 @@ class BootTimes(object):
 
     def get_last_boot_time(self):
         if self._last_boot is None:
-            self._last_boot = int(time.time() - get_uptime())
+            self._last_boot = int(time.time() - sysstats.get_uptime())
         return self._last_boot
